requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
openai
langchain
langchain_groq
python-dotenv
langgraph
streamlit